--- passenger-service/src/main/java/io/simakkoi9/passengerservice/model/Passenger.java
@@ -1,32 +1,0 @@
-package io.simakkoi9.passengerservice.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.sql.Timestamp;
-
-@Data
-@Entity
-@Table(name = "passengers")
-public class Passenger {
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "name", nullable = false, length = 255)
-    private String name;
-
-    @Column(name = "email", nullable = false, length = 255)
-    private String email;
-
-    @Column(name = "phone", nullable = false, length = 255)
-    private String phone;
-
-    @Column(name = "status", nullable = false)
-    private UserStatus status;
-
-    @Column(name = "created_at", nullable = false)
-    private Timestamp createdAt;
-
-}
