--- passenger-service/src/main/java/io/simakkoi9/passengerservice/model/UserStatus.java
@@ -1,5 +1,0 @@
-package io.simakkoi9.passengerservice.model;
-
-public enum UserStatus {
-    ACTIVE, DELETED
-}
