+++ passenger-service/src/main/java/io/simakkoi9/passengerservice/model/entity/UserStatus.java
@@ -1,0 +1,5 @@
+package io.simakkoi9.passengerservice.model.entity;
+
+public enum UserStatus {
+    ACTIVE, DELETED
+}
